Fix compat with serde_json
The previous code was relying on the accidental fact that only a single
impl of PartialEq<usize> existed. serde_json defines another impl of
PartialEq<usize>, which causes the constraint to be insufficient, which
results in a compile error when serde_json is added to a project with
Helix.

This commit forces the VALUE into a usize eagerly, avoiding this
inference issue.

Fixes #93

h/t @alyssais

Thanks to @nikomatsakis for helping me figure out the root cause.