Add Ruby Array -> vec coercion
This is essentially a rebase of #46 by @wagenet, updated for the new
APIs.

While it is not ideal to copy most of the time, it is perfectly safe.
Therefore we decide to add this in now to enable use cases that are
not otherwise possible while we work on a better non-copy version
(something like #100).

TODO: port the vec -> Ruby Array code from #46

Paired with @hone