impl `Sync` for `ID`
While we may be treating the representation of this type as opaque, I
think we can reasonably assume that it will never change to anything
that isn't `Sync`.

I think it may be worth considering making `VALUE` be `Sync` as well.
Even though it is effectively a pointer, it's a pointer to a *Ruby*
object, and therefore subject to the GVL.